Changed main.py to app.py and cli.py to main.py; main.py now has functioning argparse code for settings; Still TODO: Call main class in app.py from main.py, feed the params to class and save in settings.
